Added EFCore and implemented Expense getById and insert methods

+++ Onyx.Domain/Models/Consumer.cs
@@ -3,5 +3,5 @@ namespace Onyx.Domain.Models;
 public record Consumer
 (
     Guid UserId, 
-    decimal Share = 1
+    double Share = 1
 );
